Fix release build job
The CSS code for the `flipflop` web interface is generated
by the rake tast `assets:compile`, hence it should not
be commited to the repository at all, because its content
depends on the actual versions of gems activated at build time.
We therefore remove it and add it to `.gitignore`.

Furthermore, in #40 the stylesheet file was renamed from
`_flipflop.css` to `_flipflop.html.erb` which is problematic because

- the file is not HTML
- the file is not a template
- the filename in `rake assets:compile` was not adjusted,
  so the generated content is now ignored and instead
  the commited content is used.

We revert the renaming of the file and use `render(file: ...`
to insert its contents into the view, thereby fixing all issues.